[SYCL][Fusion][NoSTL][NFC] Move SYCLModuleInfo to its own header
